Move Theme to `:core:ui:theme:api` module
