+++ core/ui/theme/api/src/main/kotlin/app/k9mail/core/ui/theme/api/Theme.kt
@@ -1,0 +1,6 @@
+package app.k9mail.core.ui.theme.api
+
+enum class Theme {
+    LIGHT,
+    DARK,
+}
